Fix a signed/unsigned warning in RestartParticleL.

+++ src/picongpu/include/plugins/hdf5/restart/RestartParticleLoader.hpp
@@ -277,7 +277,7 @@ public:
 
             // write to frame
             assert(localId < (uint32_t)PMacc::math::CT::volume<SuperCellSize>::type::value);
-            assert((uint32_t) (localCellId) < PMacc::math::CT::volume<SuperCellSize>::type::value);
+            assert((uint32_t)localCellId < (uint32_t)PMacc::math::CT::volume<SuperCellSize>::type::value);
 
             PMACC_AUTO(particle, ((*frame)[localId]));
